# src/backend/app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Фильм на вечер API
  description: API для приложения управления коллекциями фильмов
  version: 1.0.0
servers:
  - url: /api
    description: Основной API сервер
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Movie:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Побег из Шоушенка"
        description:
          type: string
          example: "Двое заключенных находят утешение и искупление через акты порядочности..."
        image_url:
          type: string
          example: "https://example.com/movie.jpg"
        image_base64:
          type: string
          nullable: true
        rating:
          type: number
          format: float
          example: 9.3
        genre:
          type: string
          example: "Драма"
        duration:
          type: integer
          example: 142
        year:
          type: integer
          example: 1994
        loaded_from_tmdb:
          type: boolean
          example: true
      required:
        - id
        - title
    Collection:
      type: object
      properties:
        id:
          type: string
          example: "1"
        name:
          type: string
          example: "Мои любимые фильмы"
        description:
          type: string
          example: "Коллекция моих любимых фильмов"
        isDefault:
          type: boolean
          example: false
        isGroupCollection:
          type: boolean
          example: false
        groupId:
          type: string
          nullable: true
        users:
          type: array
          items:
            $ref: '#/components/schemas/CollectionUser'
        movies:
          type: array
          items:
            $ref: '#/components/schemas/MovieWithNotes'
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
      required:
        - id
        - name
        - description
        - isDefault
        - createdAt
        - updatedAt
    CollectionUser:
      type: object
      properties:
        userId:
          type: string
          example: "1"
        telegramId:
          type: string
          example: "123456789"
        username:
          type: string
          example: "user123"
        avatarUrl:
          type: string
          example: "https://example.com/avatar.jpg"
        addedAt:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
      required:
        - userId
        - telegramId
        - username
        - avatarUrl
        - addedAt
    MovieWithNotes:
      allOf:
        - $ref: '#/components/schemas/Movie'
        - type: object
          properties:
            notes:
              type: string
              nullable: true
              example: "Отличный фильм, смотрел дважды"
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 400
        message:
          type: string
          example: "Неверные данные запроса"
        details:
          type: object
          properties:
            errors:
              type: string
              example: "Отсутствует обязательное поле: name"
      required:
        - code
        - message
...
    CreateCollectionRequest:
      type: object
      properties:
        name:
          type: string
          example: "Мои любимые фильмы"
        description:
          type: string
          example: "Коллекция моих любимых фильмов"
        isDefault:
          type: boolean
          example: false
          default: false
        initialUsers:
          type: array
          items:
            $ref: '#/components/schemas/CollectionUserRequest'
          default: []
      required:
        - name
        - description
    CollectionUserRequest:
      type: object
      properties:
        telegramId:
          type: integer
          format: int64
          example: 123456789
      required:
        - telegramId
    AddMovieRequest:
      type: object
      properties:
        movieId:
          type: integer
          format: int64
          example: 1
        notes:
          type: string
          nullable: true
          example: "Отличный фильм, смотрел дважды"
      required:
        - movieId
...
security:
  - BearerAuth: []
...
paths:
...
  /collections/:
    get:
      tags:
        - Коллекции
      summary: Получить коллекции пользователя
      description: Получить все коллекции для аутентифицированного пользователя
      responses:
        '200':
          description: Список коллекций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collection'
    
    post:
      tags:
        - Коллекции
      summary: Создать новую коллекцию
      description: Создать новую коллекцию для аутентифицированного пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
      responses:
        '200':
          description: Коллекция успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /collections/{collectionId}:
    parameters:
      - name: collectionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Идентификатор коллекции
    
    get:
      tags:
        - Коллекции
      summary: Получить коллекцию по ID
      description: Получить конкретную коллекцию по её идентификатору
      responses:
        '200':
          description: Детали коллекции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Коллекция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Коллекции
      summary: Обновить коллекцию
      description: Обновить существующую коллекцию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
      responses:
        '200':
          description: Коллекция успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Коллекция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Коллекции
      summary: Удалить коллекцию
      description: Удалить коллекцию по её идентификатору
      responses:
        '200':
          description: Коллекция успешно удалена
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    example: "success"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Коллекция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  # ===== ФИЛЬМЫ В КОЛЛЕКЦИЯХ =====
  /collections/{collectionId}/movies:
    parameters:
      - name: collectionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Идентификатор коллекции
    
    post:
      tags:
        - Фильмы в коллекциях
      summary: Добавить фильм в коллекцию
      description: Добавить фильм в конкретную коллекцию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMovieRequest'
      responses:
        '200':
          description: Фильм успешно добавлен в коллекцию
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Коллекция или фильм не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /collections/{collectionId}/movies/{movieId}:
    parameters:
      - name: collectionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Идентификатор коллекции
      - name: movieId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Идентификатор фильма
    
    delete:
      tags:
        - Фильмы в коллекциях
      summary: Удалить фильм из коллекции
      description: Удалить фильм из конкретной коллекции
      responses:
        '200':
          description: Фильм успешно удален из коллекции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Коллекция или фильм не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Фильмы в коллекциях
      summary: Обновить заметки к фильму в коллекции
      description: Обновить заметки к фильму в конкретной коллекции
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                  example: "Отличный фильм, смотрел дважды"
              required:
                - notes
      responses:
        '200':
          description: Заметки к фильму успешно обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Коллекция или фильм не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /collections/{collectionId}/random:
    parameters:
      - name: collectionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Идентификатор коллекции
    
    get:
      tags:
        - Фильмы в коллекциях
      summary: Получить случайный фильм из коллекции
      description: Получить случайный фильм из конкретной коллекции
      responses:
        '200':
          description: Случайный фильм из коллекции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieWithNotes'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Коллекция не найдена или пуста
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  # ===== ПОЛЬЗОВАТЕЛИ КОЛЛЕКЦИЙ =====
  /collections/{collectionId}/users:
    parameters:
      - name: collectionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Идентификатор коллекции
    
    get:
      tags:
        - Пользователи коллекций
      summary: Получить пользователей коллекции
      description: Получить всех пользователей конкретной коллекции
      responses:
        '200':
          description: Список пользователей в коллекции
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionUser'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Коллекция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Пользователи коллекций
      summary: Добавить пользователя в коллекцию
      description: Добавить пользователя в конкретную коллекцию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUserRequest'
      responses:
        '200':
          description: Пользователь успешно добавлен в коллекцию
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Коллекция или пользователь не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /collections/{collectionId}/users/{userTelegramId}:
    parameters:
      - name: collectionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Идентификатор коллекции
      - name: userTelegramId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Telegram ID пользователя
    
    delete:
      tags:
        - Пользователи коллекций
      summary: Удалить пользователя из коллекции
      description: Удалить пользователя из конкретной коллекции
      responses:
        '200':
          description: Пользователь успешно удален из коллекции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Коллекция или пользователь не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  # ===== ПОИСК =====
...